Stemme: Fix and Re-enable with the toggle
Stemme was also not using the PlayerPrefs in
AssistanceManager. Also, when Stemme was activated
the lane was also activated. This commit ensures
that only Stemme is activated when you toggle on Stemme.
Finally, a lot of code regarding Stemme was still
present but never activated. This has been commented
out.