MainMenuOverlay: Pause the Game instead of restarting
We want to support the therapist in adapting settings
to the patient's preferences first and foremost, so it is
better if TrailiT remembers the progress made time-wise
and simple puts the level on Pause. This commit also ensures
that the level updates corresponding to how the settings change
when Resume is pressed.