UI Conversion and introduction of GameType
Introduces a Game Type toggle in Settings and makes the changes
to gameManager, AssistanceManager and LoggingManager to log it.

The toggles introduced Sliced Button Sprites which make it easier
to adhere to Stretchable UI. At some point in the future all of
the UI should be revamped to fit this stretchable paradigm.

The EndLevelCanvas was set to 800x600, this has been fixed to
2048x1536 like others and UI has been updated correspondingly.