GameManager: use Median to determine level reaction time
Averaging turns out to skew results a lot. Single values can
skew results a lot. Typically, this is for example the first
value, as we try to find number 1 and if number 1 happens
to be in the right or left, then it skews the Left/Right
reaction time averages much. Medians are more resistant
towards this as they just consider the middle value.

At the end of the session we average all medians from
each level to get a representative number.