MainMenu Buttons: Always Call GameManager.ResetGame()
Previously some of our menu buttons attempted to preserve
some state fx of the game in progress. The problem is that
we can't necessarily make much use of the results unless we
can ensure the training session has run continuously without
interruptions. This commit sanitizes all menu buttons to
call GameManager.ResetGame() which reloads the scene and
ensures a completely clean state.