Use Count Animation for our text statistics
The count animation script supports whole numbers
and decimals. Specify a duration and the script will
calculate how much to increment the numbers to
make a counting animation in the desired length. Delaying
the count animation will delay the visibility of the text.
Add the script to any game object which has a text component
to it. Call SetTargetWholeNumber or SetTargetDecimalNumber
to set targets and start the animation.