GameManager: Display End Level Reaction Times
We were never clearing LevelReactionTimesList which meant
we were taking the median across all level reaction times,
instead of the individual level. Furthermore, floating point
errors would cause our CountAnimation to be inapproximate
about the reaction time. So we set the text displayed to
be the actual count target by the end of the animation.

Additionally we were never updating LastHitTime which was
why we always had this large number inside our calculation
of the level reaction time. This has been fixed now as well.