Prepare Migration to 2018.10.04 & a Questionaire
We are soon sending out an update to Trail it on App Store. The
plan is to ask players a few extra question to collect some
basic demographic data, if they are participating in our study.
This requires version number tracking, migration functionality,
and handling new variables across loggingManager, php scripts
and mySQL. To keep things clean we are also using a new
database, which supports the newly added variables.