DataManager: Initial Creation
We are fleshing out the performance calculation and storage
to a separate class now and expanding the calculations to be
more sophisticated.

The first goal is not to visualize it, but just to log
more sophisticated data from users which we later can
analyze ourselves.

WIP.
Need to test MatchPosToCell and AddHit functions, across
different widths and heights.