Level-Auto Generation and Data Management
This commit introduces DataManager and more precise reaction time
measurement through field division. As the pre-defined levels
cannot guarantee consistent measurements across the whole screen,
we now auto-generate levels on our own. This also keeps width
consistent across difficulty. Obstacles are removed to ensure
that the reaction time changes consistently from start of level
to end of level and might become useful to make measures of
reaction time or visual search time in the future.

The introduction of DataManager means a big cleanup of GameManager
and port of LoggingManager which now have to track a lot of new
delicious variables.