GameManager: Don't measure reaction time & distance for first hit
The reaction time is imprecise because of the start animation
of the targets spawning and because using the startLevelTime
means our formula for calculating reaction time is inconsistent.

We also cannot use 0.5f 0.5f as a default for lastHitPos, since
the truth is that there was no last hit position.