GameManager: Refactor to combat -1 calculations in dataManager
From patient setting there seems to be a code path which results
in the patient's reaction time being calculated as -1 for all
levels onward. While I haven't discovered the cause of this
I am trying to refactor GameManager in an effort to ensure
this cannot happen. DataManager's addlevelData has also
had a slight refactor.