EndSessionCanvas: Heatmap
This introduces a button in the AggregatedResultsPanel which
can be used to show a heatmap. The heatmap currently count
average reaction times over 5 as poor performances and
reaction times which are twice the median reaction time
as poor performances. This is in part based on our pilot
measurements and from existing scientific litterature.
In the future we should improve this determination by
also downscaling other values once the "poorest" value
has been found so we dont end up with a completely red
canvas.