TimeVis: Only show session times from same type of training
It is hardly comparable to look at training from GameA and
then training from GameB. Neither is it comparable to look
at training on a high difficulty level with training from
a low difficulty level. So lets filter out training which
does not correspond to the current session in terms of
these variables.