GameManager: Use SessionActive variables to determine log dumping
We were using a homebrewed variable and was checking for it
wrongly in the dump data code. This meant that we were
never actually dumping our logs down and hence we lost
all logs when we failed to upload the data. Solve this
by using the sessionActive variable like we do elsewhere
in GameManager.