LoggingManager: Fix Log Dumping
Our log dumping stopped working after we postponed upload to
after ResetGame() and this meant that we lost any data which
came from iPads that didnt have any internet connection. We
fix this now by keeping logs around in a static variable across
sessions and dumping it to disk on quit.