TimeVis: Don't Assume Year 1111 for X-Axis
Our time line was off by 1 day all the time. This turned out to
be because we were defaulting to reading the day of week for
year 1111, but in fact, different dates correspond to different
days of the week depending on what year it is. This commit ensures
that the week of day is calculated properly